Updating README.md for Gradle
I'm a fan of Forge Gradle.  Turns my 13 steps into 6 steps with less
opportunity for failure.  Also, it adds the MANIFEST.MF.
Separated out the steps that need to be done once versus every time (one
of those steps is common between the two, so it looks like 7 total but
is really 6).